requirements.txt
Django==5.0.8
djangorestframework==3.15.2
django-unfold==0.40.0
environs==11.0.0
marshmallow==3.22.0
packaging==24.1
python-dotenv==1.0.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.1
pyjwt==2.9.0
django-filter==24.3